requirements.txt
# Interface
streamlit==1.44.1

# Banco de dados
sqlalchemy==2.0.40  # ORM para SQLite, mais organizado que sqlite3 puro

# Análise de dados
pandas==2.2.3

# Visualização (opcional para dashboards ou relatórios gráficos)
matplotlib==3.10.1
plotly==6.0.1

# Utilitários (para ambiente e requisições externas)
python-dotenv==1.1.0
requests==2.32.3

# Relatórios e templates (se quiser gerar HTML ou PDFs customizados)
jinja2==3.1.6

# Dev (caso use notebook para testes ou análises pontuais)
jupyter==1.0.0
notebook==7.0.3